🐛 fix incorrect resume retrieved when viewing
When viewing, the app uses finds the resume by its path. However there can be multiple resumes with the same path, and the app would take the first result, in database order, so the retrieved resume could be an incorrect one. This fixes the problem by making the app take the latest resume by version date for the given path.