🐛 fix view mode displaying an incorrect resume
The app used to use the "path" of a resume in view mode to identify a resume (instead of the UUID in edit mode). However the path was not unique and for this reason the resume shown in view mode could not be the same as the one in edit mode. This changes the view mode to use the UUID.